+++ packages/dara-core/changelog.md
@@ -6,6 +6,7 @@ title: Changelog
 
 -   Internal (JS): added `RequestExtras` context to allow injecting additional e.g. headers into requests made by Dara in different parts of the component tree
 -   Move import discovery warnings to the `--debug` logger
+-   Fixed an issue where Local storage was not being cleared between sessions. 
 
 ## 1.4.6